mcp_server/requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
pydantic==2.8.2
requests==2.32.3
python-dotenv==1.0.1
httpx==0.27.0